+++ heliopy/data/test/test_data.py
@@ -274,7 +274,6 @@ class TestWind:
         self.starttime = datetime(2010, 1, 1, 0, 0, 0)
         self.endtime = datetime(2010, 1, 1, 23, 59, 59)
 
-    '''
     def test_mfi_h0(self):
         df = wind.mfi_h0(self.starttime, self.endtime)
         check_data_output(df)
@@ -286,7 +285,6 @@ class TestWind:
     def test_threedp_pm(self):
         df = wind.threedp_pm(self.starttime, self.endtime)
         check_data_output(df)
-    '''
 
     def test_threedp_e0_emfits(self):
         df = wind.threedp_pm(self.starttime, self.endtime)
@@ -304,6 +302,10 @@ class TestWind:
 
     def test_swe_h1(self):
         df = wind.swe_h1(self.starttime, self.endtime)
+        check_data_output(df)
+
+    def test_swe_h0(self):
+        df = wind.swe_h0(self.starttime, self.endtime)
         check_data_output(df)